@@ -196,6 +196,13 @@ abstract class DataTable extends Component
     public bool $showPageLength = true;
 
     /**
+     * Suffix for the session key
+     *
+     * @var string|null
+     */
+    public ?string $sessionKeySuffix = null;
+
+    /**
      * DataTable constructor.
      *
      * @param null $id
@@ -210,12 +217,14 @@ abstract class DataTable extends Component
     /**
      * On mount handler
      *
-     * @param Builder $builder
+     * @param Builder     $builder
+     * @param string|null $sessionKeySuffix
      *
      * @throws ContainerExceptionInterface
      * @throws NotFoundExceptionInterface
+     * @return void
      */
-    public function mount($builder)
+    public function mount($builder, ?string $sessionKeySuffix = null): void
     {
         $this->resultModel = get_class($builder->getModel());
         $this->resultIds = $builder->pluck($this->prefixCol('id'))->toArray();
@@ -223,6 +232,7 @@ abstract class DataTable extends Component
         $this->resultWithoutGlobalScope = $builder->removedScopes();
 
         $this->builder = $builder;
+        $this->sessionKeySuffix = $sessionKeySuffix;
 
         $this->readSessionMetaData();
     }
@@ -392,7 +402,7 @@ abstract class DataTable extends Component
      *
      * @param string $key Column key
      */
-    public function changeSort(string $key)
+    public function changeSort(string $key): void
     {
         if ($this->sort == $key) {
             if ($this->sortDirection == 'ASC') {
@@ -409,7 +419,7 @@ abstract class DataTable extends Component
     /**
      * Reset's the pagination after changing the number of results
      */
-    public function updatingPaginate()
+    public function updatingPaginate(): void
     {
         $this->resetPage();
     }
@@ -417,7 +427,7 @@ abstract class DataTable extends Component
     /**
      * Reset's the pagination after changing the search
      */
-    public function updatingSearch()
+    public function updatingSearch(): void
     {
         $this->resetPage();
     }
@@ -506,7 +516,13 @@ CSS;
      */
     protected function getSessionMetaDataKey(): string
     {
-        return self::SESSION_KEY_META_DATA . '.' . get_class($this);
+        $sessionKey = self::SESSION_KEY_META_DATA . '.' . get_class($this);
+
+        if ($this->sessionKeySuffix !== null) {
+            $sessionKey .= '-' . $this->sessionKeySuffix;
+        }
+
+        return $sessionKey;
     }
 
     /**
