@@ -5,6 +5,7 @@ namespace Nodus\Packages\LivewireDatatables;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\ServiceProvider;
 use Livewire\Livewire;
+use Nodus\Packages\LivewireCore\Services\CspNonce;
 use Nodus\Packages\LivewireDatatables\Livewire\ConfirmModal;
 use Nodus\Packages\LivewireCore\Services\SupportsComponentAssets;
 use Nodus\Packages\LivewireDatatables\Livewire\DataTable;
@@ -51,7 +52,7 @@ class LivewireDatatablesServiceProvider extends ServiceProvider
     private function registerBladeDirectives()
     {
         Blade::directive('livewireDatatableStyles', function () {
-            return "<style " . app(\Nodus\Packages\LivewireCore\Services\CspNonce::class)->toHtml() . ">\n" . $this->styles() . "\n</style>";
+            return "<style " . app(CspNonce::class)->toHtml() . ">\n" . $this->styles() . "\n</style>";
         });
     }
 }
