@@ -1,11 +1,12 @@
 <?php
 
+namespace Nodus\Packages\LivewireDatatables\Tests\Data\Database\Migrations;
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
-{
+return new class () extends Migration {
     public function up()
     {
         Schema::create(
@@ -26,4 +27,4 @@ class CreateUsersTable extends Migration
     {
         Schema::dropIfExists('users');
     }
-}
+};
