@@ -1,11 +1,12 @@
 <?php
 
+namespace Nodus\Packages\LivewireDatatables\Tests\Data\Database\Migrations;
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreatePostsTable extends Migration
-{
+return new class () extends Migration {
     public function up()
     {
         Schema::create(
@@ -24,4 +25,4 @@ class CreatePostsTable extends Migration
     {
         Schema::dropIfExists('posts');
     }
-}
+};
